move handling exception to class and move whole class
