--- src/Handler/Table/Preview/BadExportFilterParameters.php
@@ -1,21 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Keboola\StorageDriver\BigQuery\Handler\Table\Preview;
-
-use Keboola\StorageDriver\Contract\Driver\Exception\NonRetryableExceptionInterface;
-use Keboola\StorageDriver\Shared\Driver\Exception\Exception;
-use Throwable;
-
-class BadExportFilterParameters extends Exception implements NonRetryableExceptionInterface
-{
-    public function __construct(string $message, int $code = self::ERR_VALIDATION, ?Throwable $previous = null)
-    {
-        parent::__construct(
-            $message,
-            $code,
-            $previous
-        );
-    }
-}
